pyproject.toml
[project]
name = "aluskort"
version = "0.1.0"
description = "Cloud-Neutral Security Reasoning Control Plane"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.6.0",
    "asyncpg>=0.29.0",
    "redis[hiredis]>=5.0.0",
    "qdrant-client>=1.8.0",
    "neo4j>=5.17.0",
    "PyJWT[crypto]>=2.8.0",
    "confluent-kafka>=2.3.0",
    "pyyaml>=6.0",
    "anthropic>=0.39.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "cryptography>=42.0",
]

[tool.setuptools.packages.find]
include = ["shared*", "infra*", "entity_parser*", "sentinel_adapter*", "context_gateway*", "llm_router*", "orchestrator*", "ctem_normaliser*", "atlas_detection*", "batch_scheduler*", "ops*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
